Move ValidAggregationKeys to AggregatableKeys.ts
