+++ packages/client/src/query/aggregations/AggregationsClause.ts
@@ -15,33 +15,7 @@
  */
 
 import type { ObjectOrInterfaceDefinition } from "@osdk/api";
-import type { PropertyValueWireToClient } from "@osdk/client.api";
-import type { AggregatableKeys } from "./AggregatableKeys.js";
-
-type StringAggregateOption = "approximateDistinct";
-type NumericAggregateOption =
-  | "min"
-  | "max"
-  | "sum"
-  | "avg"
-  | "approximateDistinct";
-
-type AGG_FOR_TYPE<T> = string extends T ? StringAggregateOption
-  : number extends T ? NumericAggregateOption
-  : never;
-
-export type ValidAggregationKeys<
-  Q extends ObjectOrInterfaceDefinition,
-> = keyof (
-  & {
-    [
-      KK in AggregatableKeys<Q> as `${KK & string}:${AGG_FOR_TYPE<
-        PropertyValueWireToClient[Q["properties"][KK]["type"]]
-      >}`
-    ]?: any;
-  }
-  & { $count?: any }
-);
+import type { ValidAggregationKeys } from "./AggregatableKeys.js";
 
 export type UnorderedAggregationClause<Q extends ObjectOrInterfaceDefinition> =
   { [AK in ValidAggregationKeys<Q>]?: "unordered" };
