+++ packages/client/src/query/aggregations/AggregationsResults.ts
@@ -15,14 +15,13 @@
  */
 
 import type { ObjectOrInterfaceDefinition } from "@osdk/api";
-import type { AggregatableKeys } from "./AggregatableKeys.js";
+import type { AggregatableKeys , ValidAggregationKeys } from "./AggregatableKeys.js";
 import type { AggregateOpts } from "./AggregateOpts.js";
 import type { AggregationResultsWithGroups } from "./AggregationResultsWithGroups.js";
 import type {
   AggregationCountResult,
   AggregationResultsWithoutGroups,
 } from "./AggregationResultsWithoutGroups.js";
-import type { ValidAggregationKeys } from "./AggregationsClause.js";
 
 export type AggregationsResults<
   Q extends ObjectOrInterfaceDefinition,
